Fix in organization hooks client

+++ Octokit/Clients/IOrganizationHooksClient.cs
@@ -34,7 +34,7 @@ namespace Octokit
         /// <param name="hookId">The repository's hook id</param>
         /// <remarks>See <a href="https://developer.github.com/v3/orgs/hooks/#get-single-hook">API documentation</a> for more information.</remarks>
         [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Get", Justification = "This is ok; we're matching HTTP verbs not keywords")]
-        Task<OrganizationHook> Get(string org, int hookId);
+        Task<OrganizationHook> Get(string org, long hookId);
 
         /// <summary>
         /// Creates a hook for a organization
@@ -51,7 +51,7 @@ namespace Octokit
         /// <param name="hookId">The organizations hook id</param>
         /// <param name="hook">The hook's parameters</param>
         /// <remarks>See <a href="https://developer.github.com/v3/orgs/hooks/#edit-a-hook">API documentation</a> for more information.</remarks>
-        Task<OrganizationHook> Edit(string org, int hookId, EditOrganizationHook hook);
+        Task<OrganizationHook> Edit(string org, long hookId, EditOrganizationHook hook);
 
         /// <summary>
         /// This will trigger a ping event to be sent to the hook.
@@ -59,7 +59,7 @@ namespace Octokit
         /// <param name="org">The organizations name</param>
         /// <param name="hookId">The organizations hook id</param>
         /// <remarks>See <a href="https://developer.github.com/v3/orgs/hooks/#ping-a-hook">API documentation</a> for more information.</remarks>
-        Task Ping(string org, int hookId);
+        Task Ping(string org, long hookId);
 
         /// <summary>
         /// Deletes a hook for a organization
@@ -67,6 +67,6 @@ namespace Octokit
         /// <param name="org">The organizations name</param>
         /// <param name="hookId">The organizations hook id</param>
         /// <remarks>See <a href="https://developer.github.com/v3/orgs/hooks/#delete-a-hook">API documentation</a> for more information.</remarks>
-        Task Delete(string org, int hookId);
+        Task Delete(string org, long hookId);
     }
 }
