+++ Octokit/Clients/OrganizationHooksClient.cs
@@ -52,7 +52,7 @@ namespace Octokit
         /// <returns></returns>
         /// <remarks>See <a href="https://developer.github.com/v3/orgs/hooks/#get-single-hook">API documentation</a> for more information.</remarks>
         [ManualRoute("GET", "orgs/{org}/hooks/{hook_id}")]
-        public Task<OrganizationHook> Get(string org, int hookId)
+        public Task<OrganizationHook> Get(string org, long hookId)
         {
             Ensure.ArgumentNotNullOrEmptyString(org, nameof(org));
             Ensure.ArgumentNotNull(hookId, nameof(hookId));
@@ -80,7 +80,7 @@ namespace Octokit
         /// <remarks>See <a href="https://developer.github.com/v3/orgs/hooks/#edit-a-hook">API documentation</a> for more information.</remarks>
         /// <returns></returns>
         [ManualRoute("PATCH", "orgs/{org}/hooks/{hook_id}")]
-        public Task<OrganizationHook> Edit(string org, int hookId, EditOrganizationHook hook)
+        public Task<OrganizationHook> Edit(string org, long hookId, EditOrganizationHook hook)
         {
             Ensure.ArgumentNotNullOrEmptyString(org, nameof(org));
             Ensure.ArgumentNotNull(hook, nameof(hook));
@@ -94,7 +94,7 @@ namespace Octokit
         /// <remarks>See <a href="https://developer.github.com/v3/orgs/hooks/#ping-a-hook">API documentation</a> for more information.</remarks>
         /// <returns></returns>
         [ManualRoute("POST", "orgs/{org}/hooks/{hook_id}/pings")]
-        public Task Ping(string org, int hookId)
+        public Task Ping(string org, long hookId)
         {
             Ensure.ArgumentNotNullOrEmptyString(org, nameof(org));
             Ensure.ArgumentNotNull(hookId, nameof(hookId));
@@ -107,7 +107,7 @@ namespace Octokit
         /// <remarks>See <a href="https://developer.github.com/v3/orgs/hooks/#delete-a-hook">API documentation</a> for more information.</remarks>
         /// <returns></returns>
         [ManualRoute("DELETE", "orgs/{org}/hooks/{hook_id}")]
-        public Task Delete(string org, int hookId)
+        public Task Delete(string org, long hookId)
         {
             Ensure.ArgumentNotNullOrEmptyString(org, nameof(org));
             Ensure.ArgumentNotNull(hookId, nameof(hookId));
